requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
SQLAlchemy==2.0.32
Werkzeug==3.0.3
psutil==5.9.5
gunicorn==21.2.0
pandas==2.0.3
reportlab==4.0.4